IA-984 refactor LQAS-IM Maps
