+++ plugins/polio/js/src/pages/IM/ImMap.js
@@ -5,55 +5,23 @@ import { useSafeIntl } from 'bluesquare-components';
 import { MapComponent } from '../../components/MapComponent/MapComponent';
 import { MapLegend } from '../../components/MapComponent/MapLegend';
 import { MapLegendContainer } from '../../components/MapComponent/MapLegendContainer';
-import { LqasImPopup } from '../../components/LQAS-IM/LqasImPopUp';
+import { makePopup } from '../../utils/LqasIm.tsx';
 import { LqasImMapHeader } from '../../components/LQAS-IM/LqasImMapHeader.tsx';
-import { determineStatusForDistrict, getImStatsForRound } from './utils.ts';
+import { determineStatusForDistrict, makeImMapLegendItems } from './utils.ts';
 import { districtColors } from './constants';
-import {
-    getScopeStyle,
-    findDataForShape,
-    makeLegendItem,
-} from '../../utils/index';
+import { getScopeStyle, findDataForShape } from '../../utils/index';
 import MESSAGES from '../../constants/messages';
-import { OK_COLOR, WARNING_COLOR, FAIL_COLOR } from '../../styles/constants';
-
-// Don't put it in utils to avoid circular dep
-const makePopup =
-    (imData, round, campaign = '') =>
-    shape => {
-        return (
-            <LqasImPopup
-                shape={shape}
-                data={imData}
-                round={round}
-                campaign={campaign}
-            />
-        );
-    };
 
 export const ImMap = ({ imData, shapes, round, campaign, scope }) => {
     const { formatMessage } = useSafeIntl();
     const [renderCount, setRenderCount] = useState(0);
-    const [passed, failed, disqualified] = getImStatsForRound(
+
+    const legendItems = makeImMapLegendItems(formatMessage)(
         imData,
         campaign,
         round,
     );
-    const passedLegendItem = makeLegendItem({
-        color: OK_COLOR,
-        value: passed?.length,
-        message: formatMessage(MESSAGES['1imOK']),
-    });
-    const disqualifiedLegendItem = makeLegendItem({
-        color: WARNING_COLOR,
-        value: disqualified?.length,
-        message: formatMessage(MESSAGES['2imWarning']),
-    });
-    const failedLegendItem = makeLegendItem({
-        color: FAIL_COLOR,
-        value: failed?.length,
-        message: formatMessage(MESSAGES['3imFail']),
-    });
+
     const getShapeStyles = useCallback(
         shape => {
             const status = determineStatusForDistrict(
@@ -82,11 +50,7 @@ export const ImMap = ({ imData, shapes, round, campaign, scope }) => {
                 <MapLegendContainer>
                     <MapLegend
                         title={formatMessage(MESSAGES.imResults)}
-                        legendItems={[
-                            passedLegendItem,
-                            disqualifiedLegendItem,
-                            failedLegendItem,
-                        ]}
+                        legendItems={legendItems}
                         width="lg"
                     />
                 </MapLegendContainer>
